+++ rotkehlchen/chain/ethereum/defi.py
@@ -2,9 +2,12 @@ from typing import TYPE_CHECKING, Optional
 
 from rotkehlchen.assets.asset import EthereumToken
 from rotkehlchen.constants.ethereum import (
+    FARM_ASSET_ABI,
+    YEARN_3CRV_VAULT,
     YEARN_ALINK_VAULT,
     YEARN_BCURVE_VAULT,
     YEARN_DAI_VAULT,
+    YEARN_GUSD_VAULT,
     YEARN_SRENCURVE_VAULT,
     YEARN_TUSD_VAULT,
     YEARN_USDC_VAULT,
@@ -29,7 +32,11 @@ CURVEFI_BUSDSWAP = EthereumConstants().contract('CURVEFI_BUSDSWAP')
 CURVEFI_RENSWAP = EthereumConstants().contract('CURVEFI_RENSWAP')
 CURVEFI_SRENSWAP = EthereumConstants().contract('CURVEFI_SRENSWAP')
 CURVEFI_SUSDV2SWAP = EthereumConstants().contract('CURVEFI_SUSDV2SWAP')
+CURVEFI_3POOLSWAP = EthereumConstants().contract('CURVEFI_3POOLSWAP')
+CURVEFI_GUSDC3CRVSWAP = EthereumConstants().contract('CURVEFI_GUSDC3CRVSWAP')
 YEARN_CONTROLLER = EthereumConstants().contract('YEARN_CONTROLLER')
+
+HARVEST_VAULTS = ('fUSDC', 'fUSDT', 'fDAI', 'fWETH', 'fTUSD', 'fWBTC', 'frenBTC', 'fcrvRenWBTC')
 
 
 def _handle_yearn_curve_vault(
@@ -123,6 +130,21 @@ def handle_underlying_price_yearn_vault(
     return usd_value
 
 
+def handle_underlying_price_harvest_vault(
+        ethereum: 'EthereumManager',
+        token: EthereumToken,
+        underlying_asset_price: Price,
+) -> FVal:
+    price_per_full_share = ethereum.call_contract(
+        contract_address=token.ethereum_address,
+        abi=FARM_ASSET_ABI,
+        method_name='getPricePerFullShare',
+        arguments=[],
+    )
+    usd_value = FVal(underlying_asset_price * price_per_full_share) / 10 ** token.decimals
+    return usd_value
+
+
 def handle_defi_price_query(
         ethereum: 'EthereumManager',
         token_symbol: str,
@@ -160,6 +182,14 @@ def handle_defi_price_query(
             div_decimals=36,
             asset_price=underlying_asset_price,
         )
+    elif token_symbol == 'y3Crv':
+        usd_value = _handle_yearn_curve_vault(
+            ethereum=ethereum,
+            curve_contract=CURVEFI_3POOLSWAP,
+            yearn_contract=YEARN_3CRV_VAULT,
+            div_decimals=36,
+            asset_price=ONE,  # assuming price of $1 for all stablecoins in pool
+        )
     elif token_symbol == 'yDAI+yUSDC+yUSDT+yTUSD':
         token = EthereumToken(token_symbol)
         usd_value = _handle_curvepool_price(ethereum, CURVEFI_YSWAP, token.decimals, ONE)
@@ -187,6 +217,12 @@ def handle_defi_price_query(
     elif token_symbol == 'crvPlain3andSUSD':
         token = EthereumToken(token_symbol)
         usd_value = _handle_curvepool_price(ethereum, CURVEFI_SUSDV2SWAP, token.decimals, ONE)
+    elif token_symbol == '3Crv':
+        token = EthereumToken(token_symbol)
+        usd_value = _handle_curvepool_price(ethereum, CURVEFI_3POOLSWAP, token.decimals, ONE)
+    elif token_symbol == 'gusd3CRV':
+        token = EthereumToken(token_symbol)
+        usd_value = _handle_curvepool_price(ethereum, CURVEFI_GUSDC3CRVSWAP, token.decimals, ONE)
     elif token_symbol == 'yDAI+yUSDC+yUSDT+yBUSD':
         token = EthereumToken(token_symbol)
         usd_value = _handle_curvepool_price(ethereum, CURVEFI_BUSDSWAP, token.decimals, ONE)
@@ -252,6 +288,23 @@ def handle_defi_price_query(
             contract=YEARN_TUSD_VAULT,
             div_decimals=token.decimals,
             asset_price=underlying_asset_price,
+        )
+    elif token_symbol == 'yGUSD':
+        assert underlying_asset_price
+        token = EthereumToken(token_symbol)
+        usd_value = handle_underlying_price_yearn_vault(
+            ethereum=ethereum,
+            contract=YEARN_GUSD_VAULT,
+            div_decimals=token.decimals,
+            asset_price=underlying_asset_price,
+        )
+    elif token_symbol in HARVEST_VAULTS:
+        assert underlying_asset_price
+        token = EthereumToken(token_symbol)
+        usd_value = handle_underlying_price_harvest_vault(
+            ethereum=ethereum,
+            token=token,
+            underlying_asset_price=underlying_asset_price,
         )
     else:
         return None